+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/view/model/GalleryViewModel.kt
@@ -509,6 +509,8 @@ class GalleryViewModel(
         } else {
             currentMediaRepository.update()
         }
+
+        fastScrollViewModel.updateBubbles()
     }
 
     fun loadMore() {