+++ CHANGELOG.md
@@ -13,6 +13,7 @@ In fact, such videos can be played well
 - Some rare content selection bugs
 - The "Cancel download" button in the media viewer getting stuck after clicking it
 - Crash when spamming the "Download" button in the media viewer
+- Not updating dates on the gallery scroll bar when refreshing the content
 
 ### Changed
 - Media cache is now cleared when disconnecting from the library